Desktop/DiscordBot/Observer_discord_bot_02/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
google-cloud-texttospeech